+++ admin/tabs/general/ehcache/EhCachePanel.js
@@ -9,7 +9,7 @@ import {grid, gridCountLabel} from '@xh/hoist/cmp/grid';
 import {filler} from '@xh/hoist/cmp/layout';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
 import {button, exportButton} from '@xh/hoist/desktop/cmp/button';
-import {storeFilterField} from '@xh/hoist/desktop/cmp/store';
+import {storeFilterField} from '@xh/hoist/cmp/store';
 import {Icon} from '@xh/hoist/icon';
 
 import {EhCacheModel} from './EhCacheModel';