+++ mobile/appcontainer/AppContainer.js
@@ -25,6 +25,7 @@ import {messageSource} from './MessageSource';
 import {AppContainerModel} from '@xh/hoist/appcontainer/AppContainerModel';
 
 import {tabContainerImpl} from '@xh/hoist/mobile/cmp/tab/impl/TabContainer';
+import {storeFilterFieldImpl} from '@xh/hoist/mobile/cmp/store/impl/StoreFilterField';
 import {colChooser, ColChooserModel} from '@xh/hoist/mobile/cmp/grid';
 import {installMobileImpls} from '@xh/hoist/dynamics/mobile';
 import {errorBoundary} from '@xh/hoist/core/impl';
@@ -32,6 +33,7 @@ import {useOnMount} from '@xh/hoist/utils/react';
 
 installMobileImpls({
     tabContainerImpl,
+    storeFilterFieldImpl,
     colChooser,
     ColChooserModel
 });