+++ CHANGELOG.md
@@ -13,7 +13,8 @@
 parameter will allow context menus to be specifed with a simple array in addition to the function
 specification currently supported.
 * The GridModel `defaultContextMenuTokens` array has been renamed `defaultContextMenu`.
-* Charts have been moved from `desktop/cmp/charts` to `cmp/charts`.
+* `Chart` and `ChartModel` have been moved from `desktop/cmp/charts` to `cmp/charts`.
+* `StoreFilterField` have been moved from `desktop/cmp/store` to `cmp/store`.
 
 ### 🐞 Bug Fixes